+++ app/entities/users/schemas_users.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from pydantic import EmailStr, constr
 
-from entities.users.choices import Gender
+from entities.users.choices_user import Gender
 from entities.users.models_user import _UserInSchema
