+++ app/tests/order/test_order.py
@@ -12,7 +12,7 @@ from sqlmodel.ext.asyncio.session import AsyncSession
 from entities.order.models_order import OrderModel
 from tests.utils import url_reverse
 from tests.conftest import MOCK_FREEZE_TIME, engine
-from entities.order.choices import OrderFilter, StatusOrder
+from entities.order.choices_order import OrderFilter, StatusOrder
 from entities.order.models_order import OrderInSchema
 from services.stub_init_service import T_BOOK_DATE
