+++ app/tests/order/test_order_full.py
@@ -4,7 +4,7 @@ from fastapi import status
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from tests.utils import url_reverse
-from entities.order.choices import OrderFilter
+from entities.order.choices_order import OrderFilter
 from entities.order.schemas.schema_order import OrderFullResponseSchema
 from entities.order.routers_order import ORDER_PAGE_SIZE
